Added dummy switch_args_and_call and basic infra
Added meta-programming library. For now it is header only. Some basic
CMake for the library and tests.

Added switch_args_and_call interface. It is supposed to switch the
provided arguments on and off based on the provided sequence of
Boolean flags - to add or not arguments to a list of arguments. The
provided function is called with the selected arguments.